folder and file rename and structure cleanup
