+++ src/components/PreviewOverlay.tsx
@@ -2,9 +2,9 @@ import styled from "styled-components";
 import ChevronLeftSharpIcon from "@mui/icons-material/ChevronLeftSharp";
 import ChevronRightSharpIcon from "@mui/icons-material/ChevronRightSharp";
 import { Button, IconButton } from "@material-ui/core";
-import { requestTrigger } from "../../util/unityContextActions";
-import { useSelected } from "../../util/previewOverlayFun/previewOverlayfunc";
-import { ActionData } from "../../util/types";
+import { requestTrigger } from "../util/unityContextActions";
+import { useSelected } from "../util/customHooks/previewOverlayfunc";
+import { ActionData } from "../util/types";
 
 const PreviewRoot = styled.div`
   display: absolute;
