+++ src/components/WebglBox.tsx
@@ -1,10 +1,10 @@
 import styled from "styled-components";
 import Unity from "react-unity-webgl";
 import { useEffect, useState } from "react";
-import Overlay from "../Overlay/Overlay";
-import Loading from "../Loading";
-import { unityContext, loadApparatus } from "../../util/unityContextActions";
-import { ExperienceData } from "../../util/types";
+import Overlay from "./Overlay";
+import Loading from "./Loading";
+import { unityContext, loadApparatus } from "../util/unityContextActions";
+import { ExperienceData } from "../util/types";
 
 const WebglRoot = styled.div`
   display: relative;
