+++ src/components/Workbench.tsx
@@ -3,8 +3,8 @@ import { useEffect, useState } from "react";
 import styled from "styled-components";
 import Box from "@mui/material/Box";
 import TextField from "@mui/material/TextField";
-import { getUserName } from "../../util/loginCookies";
-import { getBlobsInContainer } from "../../util/cloudOperations/readFromCloud";
+import { getUserName } from "../util/loginCookies";
+import { getBlobsInContainer } from "../util/cloudOperations/readFromCloud";
 
 const OuterBox = styled.div`
   margin-top: 5%;
