+++ src/components/editorBoxes/ActionSequenceBox.tsx
