--- src/data/fake-data-1.json
@@ -1,93 +1,0 @@
-{
-    "Id": {
-      "Type": 1,
-      "Identifier": "evil-cylinder"
-    },
-    "Transforms": [
-      {
-        "Position": {
-          "x": 0.0,
-          "y": 0.0,
-          "z": 0.0
-        },
-        "Rotation": {
-          "x": 0.0,
-          "y": 0.0,
-          "z": 0.0,
-          "w": 1.0
-        },
-        "Scale": {
-          "x": 1.0,
-          "y": 1.0,
-          "z": 1.0
-        }
-      }
-    ],
-    "Children": [
-      {
-        "Type": 0,
-        "Identifier": "evil-cylinder"
-      }
-    ],
-    "Metadata": {
-      "Paths": [
-        "evil-cylinder",
-        "evil-cylinder/evil-cylinder",
-        "evil-cylinder/evil-cylinder/delta",
-        "evil-cylinder/evil-cylinder/delta/wobble-sphere",
-        "evil-cylinder/evil-cylinder/delta/wobble-sphere/wobble-sphere",
-        "evil-cylinder/evil-cylinder/delta/wobble-sphere/wobble-sphere/animations",
-        "evil-cylinder/evil-cylinder/delta/gracious-cube",
-        "evil-cylinder/evil-cylinder/delta/earth",
-        "evil-cylinder/evil-cylinder/delta/earth/animation1",
-        "evil-cylinder/evil-cylinder/delta/gracious-cube/animation2"
-      ],
-      "Data": [
-        "0@identifier:evil-cylinder",
-        "0@type:Serialization",
-        "1@identifier:evil-cylinder",
-        "1@type:AssetBundle",
-        "1@key:evil-cylinder",
-        "2@identifier:delta",
-        "2@type:DeltaTransform",
-        "2@input:vec3/position",
-        "2@input:vec3/rotation",
-        "2@input:vec3/scale",
-        "2@input:vec3/position_delta",
-        "2@input:vec3/rotation_delta",
-        "2@input:vec3/scale_delta",
-        "2@input:bool/isLocal",
-        "3@identifier:wobble-sphere",
-        "3@type:Serialization",
-        "4@identifier:wobble-sphere",
-        "4@type:AssetBundle",
-        "4@key:wobble-sphere",
-        "5@identifier:animations",
-        "5@type:Event",
-        "5@associatedNode:wobble-sphere",
-        "5@input:void/Still",
-        "5@input:void/Wobble",
-        "5@input:void/Bounce",
-        "6@identifier:gracious-cube",
-        "6@type:AssetBundle",
-        "6@key:gracious-cube",
-        "7@identifier:earth",
-        "7@type:AssetBundle",
-        "7@key:earth",
-        "8@identifier:animations",
-        "8@type:Event",
-        "8@associatedNode:earth",
-        "8@input:void/Still",
-        "8@input:void/Wobble",
-        "9@identifier:animations",
-        "9@type:Event",
-        "9@associatedNode:gracious-cube",
-        "9@input:void/Rotate",
-        "9@input:void/Zoom-center",
-        "9@input:void/Zoom-up",
-        "9@input:void/Zoom-down",
-        "9@input:void/Rotate",
-        "9@input:void/Split"
-      ]
-    }
-  }
