--- src/data/wobble-sphere.json
@@ -1,58 +1,0 @@
-{
-    "Id": {
-      "Type": 1,
-      "Identifier": "wobble-sphere"
-    },
-    "Transforms": [
-      {
-        "Position": {
-          "x": 0.0,
-          "y": 0.0,
-          "z": 0.0
-        },
-        "Rotation": {
-          "x": 0.0,
-          "y": 0.0,
-          "z": 0.0,
-          "w": 1.0
-        },
-        "Scale": {
-          "x": 1.0,
-          "y": 1.0,
-          "z": 1.0
-        }
-      }
-    ],
-    "Children": [
-      {
-        "Type": 0,
-        "Identifier": "wobble-sphere"
-      }
-    ],
-    "Metadata": {
-      "Paths": [
-        "wobble-sphere",
-        "wobble-sphere/wobble-sphere",
-        "wobble-sphere/wobble-sphere/animations",
-        "wobble-sphere/wobble-sphere/test1",
-        "wobble-sphere/wobble-sphere/test1/test2"
-      ],
-      "Data": [
-        "0@identifier:wobble-sphere",
-        "0@type:Serialization",
-        "1@identifier:wobble-sphere",
-        "1@type:AssetBundle",
-        "1@key:wobble-sphere",
-        "2@identifier:animations",
-        "2@type:Event",
-        "2@associatedNode:wobble-sphere/wobble-sphere",
-        "2@input:void/Still",
-        "2@input:void/Wobble",
-        "2@input:void/Bounce",
-        "3@type:AssetBundle",
-        "3@identifier:test1",
-        "4@type:AssetBundle",
-        "4@identifier:test2"
-      ]
-    }
-  }
